refactor:  Separate UI State on the Screen into StateHolder and ViewModel
