+++ feature-student/outing/src/main/java/com/b1nd/dodam/outing/OutScreenState.kt
@@ -1,0 +1,41 @@
+package com.b1nd.dodam.outing
+
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.remember
+import kotlinx.datetime.toJavaLocalDateTime
+import java.time.LocalDateTime
+import java.time.temporal.ChronoUnit
+
+
+@Composable
+internal fun rememberOutScreenState(
+    lazyListState: LazyListState = rememberLazyListState(),
+): OutScreenState = remember(lazyListState) {
+    OutScreenState(lazyListState)
+}
+
+@Stable
+internal class OutScreenState(
+    val lazyListState: LazyListState,
+) {
+    private val current = LocalDateTime.now()
+
+    val canScrollBackward: Boolean
+        get() = lazyListState.canScrollBackward
+
+    fun getProgress(
+        startAt: kotlinx.datetime.LocalDateTime,
+        endAt: kotlinx.datetime.LocalDateTime
+    ): Float {
+        return ChronoUnit.SECONDS.between(
+            startAt.toJavaLocalDateTime(),
+            current,
+        ).toFloat() / ChronoUnit.SECONDS.between(
+            startAt.toJavaLocalDateTime(),
+            endAt.toJavaLocalDateTime(),
+        )
+    }
+}
