--- feature-student/outing/src/main/java/com/b1nd/dodam/outing/OutingUiState.kt
@@ -1,11 +1,0 @@
-package com.b1nd.dodam.outing
-
-import com.b1nd.dodam.data.outing.model.Outing
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
-
-data class OutingUiState(
-    val outings: ImmutableList<Outing> = persistentListOf(),
-    val sleepovers: ImmutableList<Outing> = persistentListOf(),
-    val isLoading: Boolean = false,
-)
